+++ .idea/libraries/animated_vector_drawable_25_3_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="animated-vector-drawable-25.3.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.android/build-cache/3c54d4dbdca9e05b32e3ee1b94d6987fb09c4d62/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/3c54d4dbdca9e05b32e3ee1b94d6987fb09c4d62/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/a890c5ea163925811e0627f4f8a778ee9ab1fa41/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/a890c5ea163925811e0627f4f8a778ee9ab1fa41/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
