+++ .idea/libraries/appcompat_v7_25_3_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="appcompat-v7-25.3.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.android/build-cache/ff94eb3f85515718a188cf926d07dffaf19f07d3/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/ff94eb3f85515718a188cf926d07dffaf19f07d3/output/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.android/build-cache/d7dc3e612eb833f522b21f00d2109bcceaff37a6/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/d7dc3e612eb833f522b21f00d2109bcceaff37a6/output/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
