+++ .idea/libraries/cardview_v7_25_0_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="cardview-v7-25.0.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/6f762c56bb6b85dd04ea59f41ceab49d4b885e45/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/6f762c56bb6b85dd04ea59f41ceab49d4b885e45/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/8b6402b50f2ac9f733ef8fbec15eef0c5996e46c/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/8b6402b50f2ac9f733ef8fbec15eef0c5996e46c/output/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
