+++ .idea/libraries/constraint_layout_solver_1_0_2.xml
@@ -1,7 +1,7 @@
 <component name="libraryTable">
   <library name="constraint-layout-solver-1.0.2">
     <CLASSES>
-      <root url="jar://$USER_HOME$/android-sdks/extras/m2repository/com/android/support/constraint/constraint-layout-solver/1.0.2/constraint-layout-solver-1.0.2.jar!/" />
+      <root url="jar://C:/Android/Sdk/extras/m2repository/com/android/support/constraint/constraint-layout-solver/1.0.2/constraint-layout-solver-1.0.2.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
