+++ .idea/libraries/customtabs_25_0_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="customtabs-25.0.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/2cf633008873bd2f4f6e6ee708904ec770839071/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/2cf633008873bd2f4f6e6ee708904ec770839071/output/res" />
+      <root url="file://$USER_HOME$/.android/build-cache/6ae7f9ac71909e9ecceff86b036a957af096d8f6/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/6ae7f9ac71909e9ecceff86b036a957af096d8f6/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
