+++ .idea/libraries/exposed_instrumentation_api_publish_0_5.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="exposed-instrumentation-api-publish-0.5">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/401bda41c2b69ba6b436296386337dfbb66dc99d/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/401bda41c2b69ba6b436296386337dfbb66dc99d/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/ba1a5027b82e79244c287049ebec8833c18e589a/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/ba1a5027b82e79244c287049ebec8833c18e589a/output/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
