+++ .idea/libraries/recyclerview_v7_25_3_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="recyclerview-v7-25.3.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/28cf66f455ac0617f073e8e0e3390796d413edb6/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/28cf66f455ac0617f073e8e0e3390796d413edb6/output/res" />
+      <root url="file://$USER_HOME$/.android/build-cache/08f5f125e901fc5074331034dca7b7c4c9bad2cc/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/08f5f125e901fc5074331034dca7b7c4c9bad2cc/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
