+++ .idea/libraries/support_compat_25_3_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="support-compat-25.3.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/243fcff803129053d8bd448232670907cb3ae657/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/243fcff803129053d8bd448232670907cb3ae657/output/res" />
+      <root url="file://$USER_HOME$/.android/build-cache/98776536dd3e698e5bb7e18e152bbce921e3f811/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/98776536dd3e698e5bb7e18e152bbce921e3f811/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
