+++ .idea/libraries/support_core_ui_25_3_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="support-core-ui-25.3.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/21fe1e6e56c85872996414fc651970a408294822/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/21fe1e6e56c85872996414fc651970a408294822/output/res" />
+      <root url="file://$USER_HOME$/.android/build-cache/f8f6a4fffcc532cef28ccc3693189bc1e2428464/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/f8f6a4fffcc532cef28ccc3693189bc1e2428464/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
