+++ .idea/libraries/support_core_utils_25_3_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="support-core-utils-25.3.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/7d0823ca2abac0643c7f0df2719a30acd804c8c5/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/7d0823ca2abac0643c7f0df2719a30acd804c8c5/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/b67844b6a0e9687063d8d97e730a3bd63915937b/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/b67844b6a0e9687063d8d97e730a3bd63915937b/output/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
