+++ .idea/libraries/support_fragment_25_3_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="support-fragment-25.3.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/df2b6e644eba7546d8b282936c69bd0b204e4d3a/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/df2b6e644eba7546d8b282936c69bd0b204e4d3a/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/cd5fb34b055b855c6a13ed6650e826b7ed9d7ba2/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/cd5fb34b055b855c6a13ed6650e826b7ed9d7ba2/output/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
