+++ .idea/libraries/support_media_compat_25_3_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="support-media-compat-25.3.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/6e039ee7cb74e4cb55fcedd3b3dff5604accae90/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/6e039ee7cb74e4cb55fcedd3b3dff5604accae90/output/res" />
+      <root url="file://$USER_HOME$/.android/build-cache/22463ae49ed01c8f56dc1db3b68a77bd36c36d6e/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/22463ae49ed01c8f56dc1db3b68a77bd36c36d6e/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
