+++ .idea/libraries/support_v4_25_3_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="support-v4-25.3.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/c6f5fe71deb286333ef6c6cab402eeef7595bc05/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/c6f5fe71deb286333ef6c6cab402eeef7595bc05/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/1e8620bed59cc572a75a067e636992b7272e0730/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/1e8620bed59cc572a75a067e636992b7272e0730/output/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
