+++ .idea/libraries/support_vector_drawable_25_3_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="support-vector-drawable-25.3.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.android/build-cache/dba67834e74e7e5ff22d97cd1c9091559d7db19b/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/dba67834e74e7e5ff22d97cd1c9091559d7db19b/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/eee851dc7c7ce5ad1b540218788a6076c50d4474/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/eee851dc7c7ce5ad1b540218788a6076c50d4474/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
