+++ .idea/libraries/transition_25_3_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="transition-25.3.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.android/build-cache/fc1be6aa3db3f6fae27643c5933daecc1fb820da/output/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.android/build-cache/fc1be6aa3db3f6fae27643c5933daecc1fb820da/output/res" />
+      <root url="file://$USER_HOME$/.android/build-cache/d1525ef867c27262dff05ec0a7e671992c5aebcb/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/d1525ef867c27262dff05ec0a7e671992c5aebcb/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
