+++ .idea/libraries/tweet_ui_2_3_2.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="tweet-ui-2.3.2">
     <CLASSES>
-      <root url="file://$USER_HOME$/.android/build-cache/89669bec858f3bef7887affa8c4fa16a96226774/output/res" />
-      <root url="jar://$USER_HOME$/.android/build-cache/89669bec858f3bef7887affa8c4fa16a96226774/output/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.android/build-cache/038845009cf639ca7f147a35de7a8717a20f7e39/output/res" />
+      <root url="jar://$USER_HOME$/.android/build-cache/038845009cf639ca7f147a35de7a8717a20f7e39/output/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
