+++ app/src/main/java/edu/upc/pes/wallachange/MainActivity.java
@@ -93,6 +93,10 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
                 AlertDialog dialog2 = alertBuilder2.create();
                 dialog2.show();
                 break;
+            case R.id.navigationNewItem:
+                break;
+            case R.id.navigationProfile:
+                break;
             default:
                 break;
         }
