+++ app/src/main/res/menu/drawer_activity_main_navigation.xml
@@ -9,5 +9,11 @@
         <item
             android:id="@+id/navigationLogout"
             android:title="@string/navigationLogout_eng" />
+        <item
+            android:id="@+id/navigationNewItem"
+            android:title="@string/navigationNewItem_eng" />
+        <item
+            android:id="@+id/navigationProfile"
+            android:title="@string/navigationProfile_eng" />
     </group>
 </menu>
